Menus

+++ src/main/java/com/example/finalproject/domain/menus/repository/MenuOptionsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MenuOptionsRepository extends JpaRepository<MenuOptions, Integer> {
+public interface MenuOptionsRepository extends JpaRepository<MenuOptions, Long> {
     List<MenuOptions> findByMenuId(Long menuId);
 }
